+++ libs/core/src/index.ts
@@ -2,8 +2,7 @@ export * from './lib/kondah'
 export * from './lib/types'
 export * from './lib/server-adapter'
 export * from './lib/plugin'
-export * from './lib/injectable.decorator'
-export * from './lib/add-to-context.decorator'
+export * from './lib/decorators'
 export * from './lib/metadata.types'
 
 // Re-export interface to allow declaration merging
