+++ libs/core/src/lib/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './add-to-context.decorator'
+export * from './injectable.decorator'
